requirements.txt


streamlit==1.31.0
pandas==2.0.3
openpyxl==3.1.2
pillow==10.0.0
requests==2.31.0
openai==1.5.0
python-dotenv==1.0.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.14
sounddevice==0.5.0
scipy==1.14.1
flask==3.0.3